rename battleNavActions file
