+++ modules/ui/src/contexts/index.ts
@@ -1,6 +1,6 @@
 export * from "./toaster";
 export * from "./appActions";
-export * from "./battleActions";
+export * from "./battleNavActions";
 export * from "./appState";
 export * from "./localActions";
 export * from "./gamePayload";
